Feat: 상태코드 구현

See merge request s11-webmobile1-sub2/S11P12A505!42

+++ backend/src/main/java/com/ssafy/a505/domain/service/UploadService.java
@@ -6,7 +6,8 @@ import com.ssafy.a505.domain.entity.Voice;
 
 public interface UploadService {
 
-    Voice uploadAndSendVoice(VoiceCreateRequestDTO voiceCreateRequestDTO, int pitchShift) throws JsonProcessingException;
+    Voice uploadAndSendVoice(VoiceCreateRequestDTO voiceCreateRequestDTO, float pitchShift) throws JsonProcessingException;
 
-    String sendToFlask(Voice voice, int pitchShift) throws JsonProcessingException;
+    String sendToFlask(Voice voice, float pitchShift) throws JsonProcessingException;
+
 }
